refactor(qbittorrent-e2e): extract torrent artifact builders
